serfcomm: remove commented out use of newHost
newHost was used in response to another host joining the cluster,
and sending his initial state.

In response to the initial state coming from the new joiner,
each other member of the cluster would send the full state
to the new joiner.

I am not sure this is a great idea, considering we have to
regularly send the state anyway due to updates to host
and VMs state.

Signed-off-by: Claudio Fontana <cfontana@suse.de>