pkg/model: seq replaced with ts
the problem with sequence numbers is that a libvirt restart or
an agent restart forces a new connection, and that will reset the
seq to 0, so we would then throw away all events until we reach
the same uptime as before. It is easier to use a 64bit UTC ts.

Signed-off-by: Claudio Fontana <cfontana@suse.de>