hypervisor: remove New(), replace with Connect()
There can be only one instance of hypervisor,
and the New() pattern breaks when trying to use libvirt Callbacks,
which in golang do not offer any UserData to pass along,
since libvirt golang bindings use the C API UserData for internal
purposes. So there is no way to get the instance to the callback.

Flatten out the situation and remove the anyway useless New()
instancing pattern for type Hypervisor.

Signed-off-by: Claudio Fontana <cfontana@suse.de>