virtx: make use of google UUID generator
adapt the code to explicitly generate uuids for libvirt,
do not rely on auto-generated uuids from libvirt anymore.

Use the Title field of the domain XML as the actual domain name,
for descriptive and search purposes, as the Title does not create
problems during migration or update with clashes as Name does.

Signed-off-by: Claudio Fontana <cfontana@suse.de>