virtx: http: new http_decode_body()
also new type VirtxRequest, wrapping *http.Request

we need to do this, because of forwarding (proxying).

We may need to read http.Request body multiple times,
and the fact that it is a io.ReadCloser is a problem:

r.Body can be read only once, then it needs to be closed.
So we read the r.Body fully into a []byte,
with opportune length checks,

and then whenever we need to reuse the body, we can create
a new reader to read from it multiple times.

Signed-off-by: Claudio Fontana <cfontana@suse.de>