inventory: add Hostdata with info about vms for each host
it is faster to lookup all vms contained in a host if we
maintain a VM uuid hash map for each host.

Creates a bit of a question regarding ordering of events in serf,
since if we get the VM Events before we have the host event
of the host actually owning that VM, it can be a problem.

Signed-off-by: Claudio Fontana <cfontana@suse.de>