inventory: add Host to the VmEvent
adding the hypervisor host uuid to the VmEvent
allows us to detect events that signal the change of host,

which specifically useful to capture and handle the final RESUME event
on the new host after a migration,
as well as other corner cases (missed events).

update_vm_state: handle the case of host changing.

update_vm: call update_vm_state, to be able to use a regular Vmdata refresh
also to update hosts, in case the migration event has been missed.

hypervisor: add hv.uuid to the generated VmEvent

Signed-off-by: Claudio Fontana <cfontana@suse.de>