serfcomm: fix member-failed and member-leave events handling
setting of the Tags was not inline with the reading,
so member-failed and member-leave event were never handled properly.

Change use of interface{} with "any", which is less terrible to read.

There is more to think about when it comes to give a proper meaning
to a node being 'active'. Currently this mostly reflects the serf state,
but with caveats, as a node that has failed or left won't be set to
active until virtxd sets the state to active by sending host events.

This is just how it is currently, not necessarily intentional.

Signed-off-by: Claudio Fontana <cfontana@suse.de>