virtx: change some HTTP error status codes
Instead of the generic 500, return a 424 Failed Dependency
when a libvirt operation has failed.

This helps us distinguish clearly when an error is a result of
a libvirt-related operation.

I also considered the non-standard 420 Method Failure
of the Spring framework, but 424 has the advantage of being
indeed a standard HTTP status code.

Signed-off-by: Claudio Fontana <cfontana@suse.de>