config.json: add the bind configuration to the example
It can be helpful to run serf listening only on specific interfaces,
where a specific network is used for control traffic.

By changing the config.json example to mention "bind",
the users are aware of which parameter to change if they want
to bind to a specific address.

If the user wants to bind to an interface they can do that too
with "interface", but I presume address will be the most used.

Signed-off-by: Claudio Fontana <cfontana@suse.de>