hypervisor: use flags INACTIVE, remove MIGRATABLE and SECURE
currently we use the XML only to carry the information used to
define the VM from JSON, so for "emergency" offline migrations
we do not really care if it is a MIGRATABLE configuration or not,
since we redefine the VM from scratch.

libvirt has the quirk that the MIGRATABLE flag strips the
firmware type field from the XML for inscrutable reasons,
so use INACTIVE to avoid that.

Signed-off-by: Claudio Fontana <cfontana@suse.de>