hypervisor: fix broken cpu utilization calculation
commit be7e0c47 ("hypervisor: add hugepage stats accounting")

moved code around in get_domain_stats(), moving the code that
set vm.Stats.Vcpus after the code that used it.

Rectify the situation, add some debug statements for future
improvements and reference, and while we are at it remove the
check on VM runstate before doing calculations, since otherwise
we can show an old utilization value for a long time.

Fixes: be7e0c471b83a3b39c94049dfa8b1e45cf2808fc
Signed-off-by: Claudio Fontana <cfontana@suse.de>