vmdef: remove Write_osdisk_json
this functionality is only useful for NFS virtual disks,
makes no sense for LUNs (where to put the .json?)

I never really used it, so shouldn't be a problem to get rid of this.
The user is expected to store their jsons they want to reuse
(or the json of an existing VM can be captured via get vm).

Signed-off-by: Claudio Fontana <cfontana@suse.de>